+++ src/Instance/ItemsList/ItemsList.js
@@ -19,14 +19,14 @@ import {
   MCLPagingTypes,
 } from '@folio/stripes/components';
 
-import { noValue } from '../../constants';
+import {
+  DEFAULT_ITEM_TABLE_SORTBY_FIELD,
+  noValue,
+} from '../../constants';
 import { checkIfArrayIsEmpty } from '../../utils';
 
 import ItemBarcode from './ItemBarcode';
 import ItemsListRow from './ItemsListRow';
-import {
-  sortItems,
-} from './utils';
 import useBoundWithHoldings from '../../Holding/ViewHolding/HoldingBoundWith/useBoundWithHoldings';
 
 import { DataContext } from '../../contexts';
@@ -85,7 +85,6 @@ const getFormatter = (
   },
   'status': x => x.status?.name || noValue,
   'copyNumber': ({ copyNumber }) => copyNumber || noValue,
-  'materialType': x => x.materialType?.name || noValue,
   'loanType': x => x.temporaryLoanType?.name || x.permanentLoanType?.name || noValue,
   'effectiveLocation': x => {
     const effectiveLocation = locationsById[x.effectiveLocation?.id];
@@ -100,6 +99,7 @@ const getFormatter = (
   'chronology': x => x.chronology || noValue,
   'volume': x => x.volume || noValue,
   'yearCaption': x => x.yearCaption?.join(', ') || noValue,
+  'materialType': x => x.materialType?.name || noValue,
 });
 const getColumnMapping = (intl, holdingsRecordId, items, ifItemsSelected, selectItemsForDrag) => ({
   'dnd': '',
@@ -116,13 +116,13 @@ const getColumnMapping = (intl, holdingsRecordId, items, ifItemsSelected, select
   'barcode': intl.formatMessage({ id: 'ui-inventory.item.barcode' }),
   'status': intl.formatMessage({ id: 'ui-inventory.status' }),
   'copyNumber': intl.formatMessage({ id: 'ui-inventory.copyNumber' }),
-  'materialType': intl.formatMessage({ id: 'ui-inventory.materialType' }),
   'loanType': intl.formatMessage({ id: 'ui-inventory.loanType' }),
   'effectiveLocation': intl.formatMessage({ id: 'ui-inventory.effectiveLocationShort' }),
   'enumeration': intl.formatMessage({ id: 'ui-inventory.enumeration' }),
   'chronology': intl.formatMessage({ id: 'ui-inventory.chronology' }),
   'volume': intl.formatMessage({ id: 'ui-inventory.volume' }),
   'yearCaption': intl.formatMessage({ id: 'ui-inventory.yearCaption' }),
+  'materialType': intl.formatMessage({ id: 'ui-inventory.materialType' }),
 });
 const visibleColumns = [
   'barcode',
@@ -144,19 +144,21 @@ const ItemsList = ({
   holding,
   items,
   setOffset,
+  setSorting,
   total,
   draggable,
   offset,
   isItemsDragSelected,
   selectItemsForDrag,
   getDraggingItems,
+  isFetching,
 }) => {
   const { boundWithHoldings: holdings, isLoading } = useBoundWithHoldings(items);
   const holdingsMapById = keyBy(holdings, 'id');
   const intl = useIntl();
   const [itemsSorting, setItemsSorting] = useState({
     isDesc: false,
-    column: 'barcode',
+    column: DEFAULT_ITEM_TABLE_SORTBY_FIELD,
   });
   const [records, setRecords] = useState([]);
 
@@ -185,8 +187,8 @@ const ItemsList = ({
   };
 
   useEffect(() => {
-    setRecords(checkIfArrayIsEmpty(sortItems(items, itemsSorting)));
-  }, [items, itemsSorting]);
+    setRecords(checkIfArrayIsEmpty(items));
+  }, [items]);
 
   // NOTE: in order to sort on a particular column, it must be registered
   // as a sorter in '../utils'. If it's not, there won't be any errors;
@@ -202,6 +204,7 @@ const ItemsList = ({
     };
 
     setItemsSorting(newItemsSorting);
+    setSorting(`${newItemsSorting.isDesc ? '-' : ''}${newItemsSorting.column}`);
   }, [itemsSorting]);
 
   if ((!draggable && isEmpty(items)) || isLoading) return null;
@@ -210,7 +213,7 @@ const ItemsList = ({
     <MultiColumnList
       id={`list-items-${holding.id}`}
       columnIdPrefix={`list-items-${holding.id}`}
-      contentData={items}
+      contentData={records}
       rowMetadata={rowMetadata}
       formatter={formatter}
       visibleColumns={draggable ? dragVisibleColumns : visibleColumns}
@@ -220,6 +223,7 @@ const ItemsList = ({
       onNeedMoreData={onNeedMoreData}
       pagingType={MCLPagingTypes.PREV_NEXT}
       totalCount={total}
+      nonInteractiveHeaders={['loanType', 'effectiveLocation', 'materialType']}
       onHeaderClick={onHeaderClick}
       sortDirection={itemsSorting.isDesc ? 'descending' : 'ascending'}
       sortedColumn={itemsSorting.column}
@@ -229,6 +233,7 @@ const ItemsList = ({
       pagingCanGoPrevious={pagingCanGoPrevious}
       pagingCanGoNext={pagingCanGoNext}
       hidePageIndices
+      loading={isFetching}
     />
   );
 };
@@ -237,12 +242,14 @@ ItemsList.propTypes = {
   holding: PropTypes.object.isRequired,
   items: PropTypes.arrayOf(PropTypes.object),
   setOffset: PropTypes.func.isRequired,
+  setSorting: PropTypes.func.isRequired,
   offset: PropTypes.number,
   total: PropTypes.number,
   draggable: PropTypes.bool,
   selectItemsForDrag: PropTypes.func.isRequired,
   isItemsDragSelected: PropTypes.func.isRequired,
   getDraggingItems: PropTypes.func.isRequired,
+  isFetching: PropTypes.bool,
 };
 
 ItemsList.defaultProps = {