+++ src/Instance/ItemsList/ItemsListContainer.js
@@ -1,14 +1,18 @@
-import React, { useContext, memo, useState } from 'react';
+import React, {
+  useContext,
+  memo,
+  useState,
+  useEffect,
+} from 'react';
 import PropTypes from 'prop-types';
-
-import {
-  Loading,
-} from '@folio/stripes/components';
+import { isEmpty } from 'lodash';
 
 import DnDContext from '../DnDContext';
 import ItemsList from './ItemsList';
 
 import useHoldingItemsQuery from '../../hooks/useHoldingItemsQuery';
+
+import { DEFAULT_ITEM_TABLE_SORTBY_FIELD } from '../../constants';
 
 const ItemsListContainer = ({
   holding,
@@ -24,16 +28,21 @@ const ItemsListContainer = ({
   } = useContext(DnDContext);
 
   const [offset, setOffset] = useState(0);
+  const [itemsToShow, setItemsToShow] = useState([]);
+  const [sortBy, setSortBy] = useState(DEFAULT_ITEM_TABLE_SORTBY_FIELD);
   const searchParams = {
+    sortBy,
     limit: 200,
     offset,
   };
 
   const { isFetching, items, totalRecords } = useHoldingItemsQuery(holding.id, { searchParams });
 
-  if (isFetching) {
-    return <Loading size="large" />;
-  }
+  useEffect(() => {
+    if (!isEmpty(items)) {
+      setItemsToShow(items);
+    }
+  }, [items]);
 
   return (
     <ItemsList
@@ -45,10 +54,12 @@ const ItemsListContainer = ({
       holding={holding}
       offset={offset}
       setOffset={setOffset}
-      items={items}
+      items={itemsToShow}
+      setSorting={setSortBy}
       total={totalRecords}
       draggable={draggable}
       droppable={droppable}
+      isFetching={isFetching}
     />
   );
 };